basic buddy & slab allocator

+++ include/list.h
@@ -110,7 +110,7 @@ static inline int list_empty(const struct list_head *head)
 	for (pos = (head)->next; pos != (head); pos = pos->next)
 
 
-#define offsetof(TYPE, MEMBER) ((unsigned int) &((TYPE *)0)->MEMBER)
+#define offsetof(TYPE, MEMBER) ((unsigned long) &((TYPE *)0)->MEMBER)
 
 #define container_of(ptr, type, member) ({                      \
         const typeof( ((type *)0)->member ) *__mptr = (ptr);    \
